Refactor and document auto layouting
